pyproject.toml
[project]
name = "manage_django_project"
version = "0.0.1"
description = "Helper to develop Django projects."
readme = "README.md"
authors = [
    {name = 'Jens Diemer', email = 'git@jensdiemer.de'}
]
requires-python = ">=3.9,<4"
dependencies = [
    "django-tools",  # https://github.com/jedie/django-tools
    "manageprojects",  # https://github.com/jedie/manageprojects
    "bx_py_utils",  # https://github.com/boxine/bx_py_utils
    "django-rich",  # https://github.com/adamchainz/django-rich
]
[project.optional-dependencies]
dev = [
    "pip-tools",  # https://github.com/jazzband/pip-tools/
    "tox",  # https://github.com/tox-dev/tox
    "coverage",  # https://github.com/nedbat/coveragepy
    "darker>=1.7",  # https://github.com/akaihola/darker
    "autopep8",  # https://github.com/hhatto/autopep8
    "pyupgrade",  # https://github.com/asottile/pyupgrade
    "isort",  # https://github.com/pycqa/isort
    "flynt",  # https://github.com/ikamensh/flynt
    "flake8",  # https://github.com/pycqa/flake8
    "pyflakes",  # https://github.com/PyCQA/pyflakes
    "codespell",  # https://github.com/codespell-project/codespell
    "EditorConfig",  # https://github.com/editorconfig/editorconfig-core-py
    "safety",  # https://github.com/pyupio/safety
    "mypy",  # https://github.com/python/mypy
    "twine",  # https://github.com/pypa/twine

    "tomli",  # https://github.com/hukkin/tomli
    # tomli only needed for Python <3.11, but see bug:
    # https://github.com/pypa/pip/issues/9644#issuecomment-1456583402
    #"tomli;python_version<\"3.11\"",  # https://github.com/hukkin/tomli

    # Work-a-round for:
    # https://github.com/jazzband/pip-tools/issues/994#issuecomment-1321226661
    "typing-extensions>=3.10;python_version<\"3.10\"",
]

[project.urls]
Documentation = "https://github.com/jedie/manage_django_project"
Source = "https://github.com/jedie/manage_django_project"

[project.scripts]
# Must be set in ./manage.py and PROJECT_SHELL_SCRIPT:
manage_django_project_example = "manage_django_project_example.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["manage_django_project*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[tool.darker]
src = ['.']
revision = "origin/main..."
line_length = 119
verbose = true
color = true
skip_string_normalization = true
diff = false
check = false
stdout = false
isort = true
flynt = true
lint = [
    "flake8",
]
log_level = "INFO"


[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/config_files/#pyprojecttoml-preferred-format
atomic=true
profile='black'
skip_glob=['.*', '*/htmlcov/*']
known_first_party=['manage_django_project']
line_length=119
lines_after_imports=2


[tool.coverage.run]
branch = true
parallel = true
source = ['.']
command_line = '-m manage_django_project_example test --shuffle --parallel --buffer'

[tool.coverage.report]
omit = ['.*', '*/tests/*']
skip_empty = true
fail_under = 30
show_missing = true
exclude_lines = [
    'if self.debug:',
    'pragma: no cover',
    'raise NotImplementedError',
    'if __name__ == .__main__.:',
]


[tool.tox]  # https://tox.wiki/en/latest/config.html#pyproject-toml
legacy_tox_ini = """
[tox]
isolated_build = True
envlist = py{311,310,39}
skip_missing_interpreters = True

[testenv]
passenv = *
skip_install = true
deps = pip-tools
commands_pre =
    pip-sync requirements.dev.txt
commands =
    {envpython} -m manage_django_project_example coverage --no-processing --context='{envname}'
"""


[tool.mypy]
warn_unused_configs = true
ignore_missing_imports = true
allow_redefinition = true  # https://github.com/python/mypy/issues/7165
show_error_codes = true
plugins = []
exclude = ['.venv', 'tests']


[manageprojects] # https://github.com/jedie/manageprojects
initial_revision = "04d5a25"
initial_date = 2023-03-07T16:25:36+01:00
cookiecutter_template = "https://github.com/jedie/cookiecutter_templates/"
cookiecutter_directory = "piptools-python"

[manageprojects.cookiecutter_context.cookiecutter]
full_name = "Jens Diemer"
github_username = "jedie"
author_email = "git@jensdiemer.de"
package_name = "manage_django_project"
package_version = "0.0.1"
package_description = "Helper to develop Django projects."
package_url = "https://github.com/jedie/manage_django_project"
issues_url = "https://github.com/jedie/manage_django_project/issues"
license = "GPL-3.0-or-later"
_template = "https://github.com/jedie/cookiecutter_templates/"
